+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - add `email_2fa` and `email_2fa_subject` fields to `email` page
  - add `two-factor-authentication` page
  - add `twoFactorAuth` style
+ - set column `requires_2fa` to `1` for group `admin`
 
 ### Bugfix
  - properly delete jobs for action with enabled option `Clear Scheduled Jobs for This Action`. Before it deleted all jobs based on this action, now only jobs for the current user are deleted.